feat(user_logic): refacto user
