+++ lib/features/authentication/services/auth_service.dart
