+++ lib/features/login/pages/signup_page.dart
@@ -1,8 +1,7 @@
+import 'package:amicae/features/nav%20bar/nav_bar.dart';
 import 'package:flutter/material.dart';
-import '../services/auth_service.dart';
-import 'package:amicae/features/nav%20bar/nav_bar.dart';
-import 'landing_page.dart';
-import 'package:amicae/features/nav%20bar/nav_bar.dart';
+
+import '../../authentication/services/auth_service.dart';
 
 class SignupPage extends StatefulWidget {
   const SignupPage({super.key});
