+++ lib/features/user_modules/user_errors.dart
@@ -1,0 +1,8 @@
+enum UserErrors {
+  maxChangesExceeded(
+      "The limit of changes permitted to change the parameter has been reached.");
+
+  const UserErrors(this.message);
+
+  final String message;
+}
