+++ test/user_tests/admin_test.dart
@@ -1,40 +1,61 @@
-// Importation of the test framework 
-import "package:test/test.dart";
-// importation of the module for the user abstract class
-import "../user_modules/user.dart";
-// importation of the file containing the admin class 
-import "../user_modules/admin.dart";
-// importing the exception files 
-import "../exception_modules/custom_exceptions.dart";
-// import the math library 
+// import the math library
 import "dart:math";
 
-// creation of a list of names to test the middle names 
-List <String> firstNameList = ["Christella", "Claudia", "Marie", "Carole", "Solange",
-                                "Patricia", "Claudine", "Anne Rose", "Marie Denise", "Evelyne"];
+import "package:amicae/features/user_modules/admin.dart";
+import "package:amicae/features/user_modules/user.dart";
+import "package:amicae/features/user_modules/user_errors.dart";
+import "package:test/test.dart";
 
+// creation of a list of names to test the middle names
+List<String> firstNameList = [
+  "Christella",
+  "Claudia",
+  "Marie",
+  "Carole",
+  "Solange",
+  "Patricia",
+  "Claudine",
+  "Anne Rose",
+  "Marie Denise",
+  "Evelyne"
+];
 
-List <String> middleNameList = ["Natalie", "Fabiola", "Margarette", "Junie", "Julienne",
-                                "Marie Louise", "Guerda", "Anouse", "Jesula", "Wideline"];
+List<String> middleNameList = [
+  "Natalie",
+  "Fabiola",
+  "Margarette",
+  "Junie",
+  "Julienne",
+  "Marie Louise",
+  "Guerda",
+  "Anouse",
+  "Jesula",
+  "Wideline"
+];
 
+List<String> lastNameList = [
+  "Saintil",
+  "Lindor",
+  "Civil",
+  "Exantus",
+  "Constant",
+  "Saint Fleur",
+  "Georges",
+  "Bien Aime",
+  "Gaspard",
+  "Lundy"
+];
 
-List <String> lastNameList = ["Saintil", "Lindor", "Civil", "Exantus", "Constant",
-                              "Saint Fleur", "Georges", "Bien Aime", "Gaspard",
-                              "Lundy"];
-
-// function where the tests are defined 
+// function where the tests are defined
 void main() {
-  // grouping of all the related tests together 
-  group("Admin", (){
-
-
+  // grouping of all the related tests together
+  group("Admin", () {
     test("firstName can only be changed a limited amount of times", () {
-      final admin = Admin("admin_username", 
-      "admin_password", 
-      firstName: "Meow", 
-      middleName: "Manoushka", 
-      lastName: "Woof", 
-      email: "meow.woof@amicae.com");
+      final admin = Admin("admin_username", "admin_password",
+          firstName: "Meow",
+          middleName: "Manoushka",
+          lastName: "Woof",
+          email: "meow.woof@amicae.com");
 
       // Initial change count should be 0
       expect(admin.firstName, equals("Meow"));
@@ -47,7 +68,7 @@ void main() {
         String randomName;
         do {
           randomName = firstNameList[random.nextInt(firstNameList.length)];
-        } while(randomName == admin.firstName);
+        } while (randomName == admin.firstName);
         admin.firstName = randomName;
         print(admin);
       }
@@ -55,31 +76,27 @@ void main() {
       // Verify that the number of changes is equal to the adminMaxChange
       expect(admin.firstNameChangeCounter, equals(User.maxChange));
 
-      // Change the first name one more time: with a name not in the list 
+      // Change the first name one more time: with a name not in the list
       bool exceptionThrown = false;
       try {
         admin.firstName = "Vanessa";
-      }
-      catch (error) {
-        if (error is MaxChangeExceeded) {
-          exceptionThrown = true;
-        }
+      } on UserErrors {
+        exceptionThrown = true;
       }
 
-      // verify that the exception was thrown 
+      // verify that the exception was thrown
       expect(exceptionThrown, isTrue);
 
-      // verify that the first name is still in the list 
+      // verify that the first name is still in the list
       expect(admin.firstName, isIn(firstNameList));
     });
 
     test("middleName can only be changed a limited amount of times", () {
-      final admin = Admin("admin_username", 
-      "admin_password", 
-      firstName: "Meow", 
-      middleName: "Manoushka", 
-      lastName: "Woof", 
-      email: "meow.woof@amicae.com");
+      final admin = Admin("admin_username", "admin_password",
+          firstName: "Meow",
+          middleName: "Manoushka",
+          lastName: "Woof",
+          email: "meow.woof@amicae.com");
 
       // Initial change count should be 0
       expect(admin.middleName, equals("Manoushka"));
@@ -92,7 +109,7 @@ void main() {
         String randomName;
         do {
           randomName = middleNameList[random.nextInt(middleNameList.length)];
-        } while(randomName == admin.middleName);
+        } while (randomName == admin.middleName);
         admin.middleName = randomName;
       }
 
@@ -103,27 +120,23 @@ void main() {
       bool exceptionThrown = false;
       try {
         admin.middleName = "Vanessa";
-      }
-      catch (error) {
-        if (error is MaxChangeExceeded) {
-          exceptionThrown = true;
-        }
+      } on UserErrors {
+        exceptionThrown = true;
       }
 
-      // verify that the exception was thrown 
+      // verify that the exception was thrown
       expect(exceptionThrown, isTrue);
 
-      // verify that the middle name is still in the list 
+      // verify that the middle name is still in the list
       expect(admin.middleName, isIn(middleNameList));
     });
 
     test("lastName can only be changed a limited amount of times", () {
-      final admin = Admin("admin_username", 
-      "admin_password", 
-      firstName: "Meow", 
-      middleName: "Manoushka", 
-      lastName: "Woof", 
-      email: "meow.woof@amicae.com");
+      final admin = Admin("admin_username", "admin_password",
+          firstName: "Meow",
+          middleName: "Manoushka",
+          lastName: "Woof",
+          email: "meow.woof@amicae.com");
 
       // Initial change count should be 0
       expect(admin.lastName, equals("Woof"));
@@ -136,7 +149,7 @@ void main() {
         String randomName;
         do {
           randomName = lastNameList[random.nextInt(lastNameList.length)];
-        } while(randomName == admin.lastName);
+        } while (randomName == admin.lastName);
         admin.lastName = randomName;
       }
 
@@ -147,107 +160,98 @@ void main() {
       bool exceptionThrown = false;
       try {
         admin.lastName = "Michel";
-      }
-      catch (error) {
-        if (error is MaxChangeExceeded) {
-          exceptionThrown = true;
-        }
+      } on UserErrors {
+        exceptionThrown = true;
       }
 
-      // verify that the exception was thrown 
+      // verify that the exception was thrown
       expect(exceptionThrown, isTrue);
 
-      // verify that the last name is still in the list 
+      // verify that the last name is still in the list
       expect(admin.lastName, isIn(lastNameList));
     });
 
+    test(
+        "changing to the same first name does not decrement the change counter",
+        () {
+      final admin = Admin("admin_username", "admin_password",
+          firstName: "Meow",
+          middleName: "Manoushka",
+          lastName: "Woof",
+          email: "meow.woof@amicae.com");
 
-    test("changing to the same first name does not decrement the change counter", () {
-      final admin = Admin("admin_username", 
-      "admin_password", 
-      firstName: "Meow", 
-      middleName: "Manoushka", 
-      lastName: "Woof", 
-      email: "meow.woof@amicae.com");
-
-      // Initial change count should be 0 
+      // Initial change count should be 0
       expect(admin.firstName, equals("Meow"));
       expect(admin.firstNameChangeCounter, equals(0));
       print(admin);
 
-      // Changing the firstName to a different name 
+      // Changing the firstName to a different name
       admin.firstName = firstNameList[0];
       expect(admin.firstName, equals(firstNameList[0]));
       expect(admin.firstNameChangeCounter, equals(1));
       print(admin);
 
-      // Changing the firstName to the same name 
+      // Changing the firstName to the same name
       admin.firstName = firstNameList[0];
       expect(admin.firstName, equals(firstNameList[0]));
       expect(admin.firstNameChangeCounter, equals(1));
       print(admin);
-        // Counter should not increment 
-
+      // Counter should not increment
     });
 
+    test(
+        "changing to the same middle name does not decrement the change counter",
+        () {
+      final admin = Admin("admin_username", "admin_password",
+          firstName: "Meow",
+          middleName: "Manoushka",
+          lastName: "Woof",
+          email: "meow.woof@amicae.com");
 
-    test("changing to the same middle name does not decrement the change counter", () {
-      final admin = Admin("admin_username", 
-      "admin_password", 
-      firstName: "Meow", 
-      middleName: "Manoushka", 
-      lastName: "Woof", 
-      email: "meow.woof@amicae.com");
-
-      // Initial change count should be 0 
+      // Initial change count should be 0
       expect(admin.middleName, equals("Manoushka"));
       expect(admin.middleNameChangeCounter, equals(0));
-      print(admin);
+      // print(admin);
 
-      // Changing the firstName to a different name 
+      // Changing the firstName to a different name
       admin.middleName = middleNameList[0];
       expect(admin.middleName, equals(middleNameList[0]));
       expect(admin.middleNameChangeCounter, equals(1));
-      print(admin);
+      // print(admin);
 
-      // Changing the firstName to the same name 
+      // Changing the firstName to the same name
       admin.middleName = middleNameList[0];
       expect(admin.middleName, equals(middleNameList[0]));
       expect(admin.middleNameChangeCounter, equals(1));
-        // Counter should not increment 
-      print(admin);
-
+      // Counter should not increment
+      // print(admin);
     });
 
+    test("changing to the same last name does not decrement the change counter",
+        () {
+      final admin = Admin("admin_username", "admin_password",
+          firstName: "Meow",
+          middleName: "Manoushka",
+          lastName: "Woof",
+          email: "meow.woof@amicae.com");
 
-    test("changing to the same last name does not decrement the change counter", () {
-      final admin = Admin("admin_username", 
-      "admin_password", 
-      firstName: "Meow", 
-      middleName: "Manoushka", 
-      lastName: "Woof", 
-      email: "meow.woof@amicae.com");
-
-      // Initial change count should be 0 
+      // Initial change count should be 0
       expect(admin.lastName, equals("Woof"));
       expect(admin.lastNameChangeCounter, equals(0));
       print(admin);
 
-      // Changing the firstName to a different name 
+      // Changing the firstName to a different name
       admin.lastName = lastNameList[0];
       expect(admin.lastName, equals(lastNameList[0]));
       expect(admin.lastNameChangeCounter, equals(1));
       print(admin);
 
-      // Changing the firstName to the same name 
+      // Changing the firstName to the same name
       admin.lastName = lastNameList[0];
       expect(admin.lastName, equals(lastNameList[0]));
       expect(admin.lastNameChangeCounter, equals(1));
-        // Counter should not increment 
+      // Counter should not increment
       print(admin);
-
     });
-
   });
-
 }
